scripts: Add options to only run parts of the index.py file
This commit only inserts the embeddings in the database if the input
file is given. This allows users to use the index.py file to load the
computed centroids even if the embeddings are already in the database.

Also, add a --noindex parameter to run the load the data without
building the index.

Lastly, remove the dbname="postgres" line to use the database name given
in the url argument.

Signed-off-by: Maxime Schoemans <maxime.schoemans@enterprisedb.com>